+++ Config/ProjectSettings.ps1
@@ -49,6 +49,7 @@ Set-Variable -Name SettingsScript -Scope Local -Option ReadOnly -Value ($MyInvoc
 # 3. Performs additional requirements checks needed or recommended for development
 # 4. Enables some disabled unit tests and disables logging
 # 5. Enables setting preference variables for modules
+# NOTE: Changing variable requires PowerShell restart
 Set-Variable -Name Develop -Scope Global -Value $true
 
 if ($Develop)
@@ -215,13 +216,13 @@ if (!(Get-Variable -Name CheckProjectConstants -Scope Global -ErrorAction Ignore
 	New-Variable -Name RequirePackageManagementVersion -Scope Global -Option Constant -Value ([version]::new(1, 4, 7))
 
 	# Required minimum PowerShellGet version prior to installing other modules, do not decrement!
-	New-Variable -Name RequirePowerShellGetVersion -Scope Global -Option Constant -Value ([version]::new(2, 2, 4))
+	New-Variable -Name RequirePowerShellGetVersion -Scope Global -Option Constant -Value ([version]::new(2, 2, 5))
 
 	# Recommended minimum platyPS version used to generate online help files for modules, do not decrement!
 	New-Variable -Name RequirePlatyPSVersion -Scope Global -Option Constant -Value ([version]::new(0, 14, 0))
 
 	# Recommended minimum VSCode version, do not decrement!
-	New-Variable -Name RequireVSCodeVersion -Scope Global -Option Constant -Value ([version]::new(1, 48, 2))
+	New-Variable -Name RequireVSCodeVersion -Scope Global -Option Constant -Value ([version]::new(1, 49, 2))
 
 	# Recommended minimum PSReadline version for command line editing experience of PowerShell
 	# Needs the 1.6.0 or a higher version of PowerShellGet to install the latest prerelease version of PSReadLine